pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.51.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.1.0",
    "slack-bolt>=1.23.0",
]
